pyproject.toml
[tool.poetry]
name = "poop_garden_api"
version = "0.1.0"
description = ""
authors = ["mhendricks96 <mhendricks96@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
django-cors-headers = "^3.11.0"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
whitenoise = "^6.0.0"
django-friendship = "^1.9.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
